Fix compiled functions when # of upvalues >= 128

+++ luaj-vm/src/core/org/luaj/compiler/FuncState.java
@@ -40,8 +40,8 @@ import org.luaj.vm.Platform;
 
 public class FuncState extends LuaC {
 	class upvaldesc {
-		  byte k;
-		  byte info;
+		  short k;
+		  short info;
 	};
 
 	static class BlockCnt {
